pyproject.toml
[tool.poetry]
name = "stock-analysis"
version = "0.1.0"
description = ""
authors = ["venkat sundaram"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
yfinance = "^0.2.31"
openpyxl = "^3.1.2"
arch = "^6.2.0"
pandas-market-calendars = "^4.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
